Switch to plain JUnit4 test suite
Add junit-interface to make sure "sbt test" does the right thing.
The drawback of JUnitSuite was that when you import to IDEA and
right click "Run" on a test class it presents two menu items that
both run the same tests which is silly (one of them runs via JUnit
and the other menu item runs the same tests but via ScalaTest
instead).