Use all CPU cores when hashing advent coins for day 4
This just makes it run in ~80% of the single core time because the
JVM eats so much time for JIT. If you run solve() a few times before
you measure, then the parallel version takes ~54% of the single core
speed instead. Better but Amdahl is still nowhere to be seen (I'm
measuring this on a quad core machine). I like how this particular
way of parallelizing causes minimal disruption to code readability
though.